Fixed halloween colors
